% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{add_scores}
\alias{add_scores}
\title{Function for adding cscores to an Upstream DB using a list with database and rank columns}
\usage{
add_scores(
  db,
  df_scores,
  add_by = c(Database = "database", Kinase_Rank = "rank")
)
}
\description{
Function for adding cscores to an Upstream DB using a list with database and rank columns
}
