% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_analysis.R
\name{addDbWeights}
\alias{addDbWeights}
\title{Add database weights to upstream kinase database
Helper function to add database-specific weights to the kinase database}
\usage{
addDbWeights(db, dbWeights)
}
\arguments{
\item{db}{data frame with upstream kinase database}

\item{dbWeights}{named vector of weights for each database, must match the Database column in db}
}
\value{
the input database with dbWeight column added
}
\description{
Add database weights to upstream kinase database
Helper function to add database-specific weights to the kinase database
}
