% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{cscore0}
\alias{cscore0}
\title{Function for simple c-scoring on standard upstream DB}
\usage{
cscore0(
  db,
  rank_table = data.frame(rank = c(0, 1, 6, 12), score = c(0.9, 0.7, 0.5, 0))
)
}
\description{
Function for simple c-scoring on standard upstream DB
}
