% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{cscore2w}
\alias{cscore2w}
\title{Function for calculating (normalized) weights based on c-scores}
\usage{
cscore2w(
  dbframe,
  sub.var = "ID",
  up.var = "Kinase_Name",
  value.var = "cscore",
  na.wt = 0
)
}
\description{
Function for calculating (normalized) weights based on c-scores
}
