% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinase_tests.R
\name{gtUKA}
\alias{gtUKA}
\title{Function for performing two group UKA test using Goeman's global test package}
\usage{
gtUKA(
  X,
  grp,
  ksr_table,
  kinase_id = "Kinase_Name",
  substarte_id = "ID",
  weight_id = "wn",
  scale = TRUE,
  directional = 1
)
}
\arguments{
\item{X}{A matrix of substrate measurements, with substrates as columns and samples as rows}

\item{grp}{A vector of group labels for the samples, must have exactly two levels}

\item{ksr_table}{A data frame containing the kinase-substrate relationships and weights.
The data frame must contain columns for kinase name, substrate ID and weight, which can be specified using the kinase_id, substrate_id and weight_id parameters}

\item{kinase_id}{The name of the column in ksr_table that contains the kinase}

\item{substarte_id}{The name of the column in ksr_table that contains the substrate ID, which should match the column names in X}

\item{weight_id}{The name of the column in ksr_table that contains the weights}

\item{scale}{Whether to standardize the substrate measurements before performing the test}

\item{directional}{Whether to perform a directional test (1) or a non-directional}
}
\description{
The function takes a kinase-substrate relationship table and performs the global test for each kinase, using the substrates as features and the group labels as response.
The weights for the substrates can be specified in the ksr_table
The function returns a data frame with the kinase name, p-value and number of substrates used in the test
}
