% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_analysis.R
\name{intersectById}
\alias{intersectById}
\title{Find intersection of peptide IDs between two data frames
Helper function to intersect data frames by ID column and ensure matching peptide IDs}
\usage{
intersectById(df1, df2)
}
\arguments{
\item{df1}{first data frame with ID column}

\item{df2}{second data frame with ID column}
}
\value{
a list containing two data frames filtered to common IDs
}
\description{
Find intersection of peptide IDs between two data frames
Helper function to intersect data frames by ID column and ensure matching peptide IDs
}
