% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_analysis.R
\name{pgCscoreAnalysis}
\alias{pgCscoreAnalysis}
\title{Function for Confidence score based UKA
Performs UKA for both comparison and fold-change analysis types}
\usage{
pgCscoreAnalysis(
  df,
  dbFrame,
  scores_df,
  nPermutations = 500,
  norm_func = "log10",
  ukaType
)
}
\arguments{
\item{df}{long format data frame with columns ID, colSeq, value, and grp (latter only for comparison mode)}

\item{dbFrame}{data frame with the UKA database}

\item{scores_df}{data frame with C-scores for normalization, containing columns Database, Kinase_Rank, cscore}

\item{nPermutations}{number of permutations to use, default 500}

\item{norm_func}{normalization function, default "log10"}

\item{ukaType}{analysis type, either "Comparison" for two-group comparison or "FC" for fold-change analysis}
}
\value{
a data frame with UKA results ordered by combined score
}
\description{
Function for Confidence score based UKA
Performs UKA for both comparison and fold-change analysis types
}
