% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream_analysis.R
\name{pgScanAnalysis2g}
\alias{pgScanAnalysis2g}
\title{perform a UKA analysis while scanning the PNET rank of upstream kinases
Intended for use with BN upstream apps
pgScanAnalysis2g(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis using binary grouping based on the grp column in df, implements parallel processing
pgScanAnalysis0(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis based on a single column (LFC) in df, no grouping, implements parallel processing
pgScanAnalysis2g_np(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis using binary grouping based on the grp column in df, no parallel processing
pgScanAnalysis0_np(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis based on a single column (LFC) in df, no grouping, no parallel processing}
\usage{
pgScanAnalysis2g(
  df,
  dbFrame,
  dbWeights,
  scanRank,
  nPermutations = 500,
  bootstrap = FALSE
)
}
\arguments{
\item{df}{long format data frame with columns ID, colSeq, value, grp}

\item{dbFrame}{data frame with the UK database}

\item{dbWeights}{named vector of weights for each database, must match the names in dbFrame}

\item{scanRank}{vector of ranks to scan}

\item{nPermutations}{number of permutations to use}

\item{bootstrap}{logical, if TRUE, a bootstrap sample of the observations is used.}
}
\value{
a list of data frames, one for each scanRank
}
\description{
perform a UKA analysis while scanning the PNET rank of upstream kinases
Intended for use with BN upstream apps
pgScanAnalysis2g(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis using binary grouping based on the grp column in df, implements parallel processing
pgScanAnalysis0(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis based on a single column (LFC) in df, no grouping, implements parallel processing
pgScanAnalysis2g_np(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis using binary grouping based on the grp column in df, no parallel processing
pgScanAnalysis0_np(df, dbFrame, dbWeights, scanRank, nPermutations = 500) UKA analysis based on a single column (LFC) in df, no grouping, no parallel processing
}
