% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{rank2cscore}
\alias{rank2cscore}
\title{Function for converting a rank metric to a c-score based on rank_table}
\usage{
rank2cscore(rank, rank_table)
}
\description{
Function for converting a rank metric to a c-score based on rank_table
}
