% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{add_scores}
\alias{add_scores}
\title{Function for adding cscores to an Upstream DB using a list with database and rank columns}
\usage{
add_scores(db, scores_df, add_by = c("Database", "Kinase_Rank"))
}
\description{
Function for adding cscores to an Upstream DB using a list with database and rank columns
}
